backend/requirements.txt
Flask
docker
flask-cors
gunicorn
psutil
bcrypt
PyJWT